refactor: Add UserSlice and configure user state in Store
This commit adds a new UserSlice file that contains the Redux slice for managing user-related state. It includes the definition of the UserState interface, initial state, and actions for setting the current user, clearing the current user, adding a user, and removing a user. The UserSlice is then imported and added to the store configuration in the Store file. This update allows for managing user data in the application's Redux store.